+++ state/pgstatestorage_test.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/0xPolygonHermez/zkevm-node/hex"
+	"github.com/0xPolygonHermez/zkevm-node/log"
 	"github.com/0xPolygonHermez/zkevm-node/state"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -436,4 +437,27 @@ func TestVirtualBatch(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, virtualBatch, *actualVirtualBatch)
 	require.NoError(t, dbTx.Commit(ctx))
+}
+
+func TestStoreEvent(t *testing.T) {
+	var readEvent state.Event
+	var hashStr string
+
+	event := &state.Event{
+		EventType: state.EventType_Prexecution_OOC,
+		Timestamp: time.Now(),
+		IP:        "127.0.0.1",
+		TxHash:    common.Hash{},
+		Payload:   "This is a test 123",
+	}
+
+	err := testState.AddEvent(ctx, event, nil)
+	if err != nil {
+		log.Errorf("Error adding event: %v", err)
+	}
+
+	err = testState.PostgresStorage.QueryRow(ctx, "SELECT event_type, timestamp, ip, tx_hash, payload FROM state.event limit 1").Scan(&readEvent.EventType, &readEvent.Timestamp, &readEvent.IP, &hashStr, &readEvent.Payload)
+	assert.NoError(t, err)
+	readEvent.TxHash = common.HexToHash(hashStr)
+	assert.Equal(t, event.Payload, readEvent.Payload)
 }
